# docs/swagger.yaml
basePath: /
definitions:
  ErrorResponse:
    properties:
      message:
        type: string
      param:
        type: string
      value:
        type: string
    type: object
...
  models.Match:
    properties:
      bookingId:
        type: integer
      cost:
        type: number
      created:
        type: string
      isPaid:
        type: boolean
      lastUpdated:
        type: string
      matchAccessStatusId:
        type: integer
      matchDate:
        type: string
      matchId:
        type: integer
      matchStatusId:
        type: integer
    required:
    - bookingId
    - cost
    - created
    - isPaid
    - lastUpdated
    - matchAccessStatusId
    - matchDate
    - matchStatusId
    type: object
...
  models.Player:
    properties:
      address:
        maxLength: 15
        minLength: 11
        type: string
      city:
        maxLength: 50
        minLength: 3
        type: string
      nickname:
        maxLength: 16
        minLength: 2
        type: string
      playerId:
        type: integer
      postcode:
        maxLength: 8
        minLength: 5
        type: string
      userId:
        type: integer
    required:
    - address
    - city
    - nickname
    - postcode
    type: object
  models.Squad:
    properties:
      city:
        type: string
      name:
        type: string
      squadId:
        type: integer
    type: object
...
  payloads.JoinMatchResponse:
    properties:
      matchId:
        type: integer
    type: object
...
  payloads.PlayerPaymentRequest:
    properties:
      amountToPay:
        type: number
    required:
    - amountToPay
    type: object
  payloads.PlayerRequest:
    properties:
      city:
        maxLength: 50
        minLength: 3
        type: string
      nickname:
        maxLength: 16
        minLength: 2
        type: string
      phoneNo:
        maxLength: 15
        minLength: 11
        type: string
      postcode:
        maxLength: 8
        minLength: 5
        type: string
    required:
    - city
    - nickname
    - phoneNo
    - postcode
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: nathan.mcalinden@unosquare.com
    name: Footpal Support
  description: This is swagger for Footpal
  termsOfService: http://swagger.io/terms/
  title: Footpal API
  version: "1.0"
paths:
...
  /players:
    get:
      description: Retrieve all players
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Player'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - player
    put:
      description: Edit player information
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.PlayerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Player'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - player
  /players/{playerId}:
    get:
      description: Get Player information
      parameters:
      - description: Player ID
        in: path
        name: playerId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Player'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - player
  /players/matches:
    get:
      description: Retrieve a players opted in matches
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Match'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - player
  /players/matches/{matchId}:
    delete:
      description: Leave match
      parameters:
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - player
    post:
      description: Request to join match
      parameters:
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payloads.JoinMatchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - player
  /players/matches/{matchId}/pay:
    post:
      description: For a match, pay amount owed by player
      parameters:
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: integer
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.PlayerPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - player
    put:
      description: For a match, pay amount owed by player
      parameters:
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: integer
      - description: Payment Type ID
        in: query
        name: paymentTypeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - player
  /players/squads:
    get:
      description: Retrieve all squads linked to player
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Squad'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - player
  /players/squads/{squadId}:
    get:
      description: Retrieve squad by squad id
      parameters:
      - description: Player ID
        in: path
        name: playerId
        required: true
        type: integer
      - description: Squad ID
        in: path
        name: squadId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Player'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - player
    post:
      description: Request to join squad
      parameters:
      - description: Squad ID
        in: path
        name: squadId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - player
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"